+++ ChangeLog.md
@@ -1,3 +1,4 @@
-- You can now expand transit lines to view their stops.
+- You can now expand transit lines to view their stops. Each stop will show connecting lines, as well as nearby stops of any type.
 - Added Busiest Stops list (click the hourglass icon).
 - Fixed scrolling when jumping to a transit line.
+- Updated subway icons to better differentiate from trains.